More Fun!
git-svn-id: http://svn.terracotta.org/svn/ehcache@6642 b9324663-ca0f-0410-8574-be9b3887307d

+++ branches/ehcache-statistics-rework/ehcache-core/src/main/java/net/sf/ehcache/hibernate/management/impl/EhcacheStatsImpl.java
@@ -326,8 +326,8 @@ public class EhcacheStatsImpl extends BaseEmitterBean implements EhcacheStats {
             Cache cache = cacheManager.getCache(name);
             if (cache != null) {
                 rv.put(name, new int[] {(int) cache.getStatistics().getExtended().getCacheHitMostRecentSample(),
-                        cache.getStatistics().getExtended().getCacheMissNotFoundMostRecentSample()
-                                + cache.getStatistics().getExtended().getCacheMissExpiredMostRecentSample(),
+                        (int)cache.getStatistics().getExtended().getCacheMissNotFoundMostRecentSample()
+                                + (int)cache.getStatistics().getExtended().getCacheMissExpiredMostRecentSample(),
                         (int) cache.getStatistics().getExtended().getCacheElementPutMostRecentSample(), });
             }
         }
